:sparkles: Migration to FastAPI
migration to fastAPI
