+++ app/main.py
@@ -4,5 +4,5 @@ from app import config as settings
 from app.api.v1.endpoint import router
 
 app = FastAPI(title=settings.PROJECT_NAME)
-
 app.include_router(router, prefix=settings.API_V1_STR)
+print("ready")
